Initial commit of slideshow command from .bash_history
Jacob got this command working, but never felt the need to put it
in a script or document it. This command works, but leaves a lot
to be desired. Must be manually run whenever the pi reboots.

I have put some basic requirements in the slideshow script.
All this stuff should be set up via ansible or similar.